% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_expression_RNA.r
\name{average_expression_RNA}
\alias{average_expression_RNA}
\title{Creates average expression matrix for a gene of interest and a list of genes}
\usage{
average_expression_RNA(obj, goi, gene_list)
}
\arguments{
\item{obj}{Seurat object. gene_list Gene list. genecolumn Gene column name}
}
\value{
A vector of present genes
}
\description{
Creates average expression matrix for a gene of interest and a list of genes
}
\examples{
average_expression_RNA(obj, goi, gene_list)


}
