% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_expression_RNA.r
\name{celltype_expression_RNA}
\alias{celltype_expression_RNA}
\title{Creates expression matrix for a gene of interest and a list of genes and a celltype}
\usage{
celltype_expression_RNA(obj, celltype, GOI, gene_list)
}
\arguments{
\item{obj}{Seurat object. goi Gene of interest. gene_list Gene list. celltype Celltype of interest}
}
\value{
A vector of present genes
}
\description{
Creates expression matrix for a gene of interest and a list of genes and a celltype
}
\examples{
celltype_expression_RNA(obj, celltype, GOI, gene_list)


}
