% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.r
\name{correlation}
\alias{correlation}
\title{Finds correlation between a gene of interest and all other genes in a Seurat object}
\usage{
correlation(df, goi, gene_list, PorS, ensembl)
}
\arguments{
\item{df}{Matrix of gene expression values}

\item{goi}{Gene of interest.}

\item{gene_list}{Gene list.}

\item{PorS}{Type of correlation to compute (P for Pearson, S for Spearman)}

\item{ensembl}{Boolean, whether to convert Ensembl IDs to gene symbols}
}
\value{
A matrix of correlation values
}
\description{
Finds correlation between a gene of interest and all other genes in a Seurat object
}
\examples{
correlation(df, goi, gene_list, PorS, ensembl)
correlation(matrix,"RFX3", gene_list, "P", ensembl = "FALSE")


}
