% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Autogenerate_help.R
\docType{package}
\name{CorrelationGene}
\alias{CorrelationGene-package}
\alias{CorrelationGene}
\title{CorrelationGene Package}
\description{
Using these functions, you can easily see the correlations when inputting a gene of interest, a gene list and a Seurat object. Please note that currently it can only be a Seurat object.
}
\author{
Timothy Allen \email{timothy.james.allen@ki.se}
}
