% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/express_average_SCT.r
\name{express_average_SCT}
\alias{express_average_SCT}
\title{Finds correlation between a gene of interest and all other genes across every level in a Seurat Obj}
\usage{
express_average_SCT(obj, goi, gene_list)
}
\arguments{
\item{obj}{Seurat object. gene_list Gene list. goi Gene of interest.}
}
\value{
data frame with correlation values
}
\description{
Finds correlation between a gene of interest and all other genes across every level in a Seurat Obj
}
\examples{
express_average_SCT(obj, goi, gene_list)


}
