% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/express_cell_RNA.r
\name{express_cell_RNA}
\alias{express_cell_RNA}
\title{Finds correlation between a gene of interest and all other genes in a Seurat object per level}
\usage{
express_cell_RNA(obj, goi, gene_list, PorS)
}
\arguments{
\item{obj}{Seurat object. gene_list Gene list. goi Gene of interest.}
}
\value{
data frame with correlation values
}
\description{
Finds correlation between a gene of interest and all other genes in a Seurat object per level
}
\examples{
express_cell_RNA(obj, goi, gene_list)


}
