% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present_genes_SCT.r
\name{present_genes_SCT}
\alias{present_genes_SCT}
\title{Find Present genes in the seurat object and the gene list}
\usage{
present_genes_SCT(obj, gene_list, genecolumn)
}
\arguments{
\item{obj}{Seurat object. gene_list Gene list. genecolumn Gene column name}
}
\value{
A vector of present genes
}
\description{
Find Present genes in the seurat object and the gene list
}
\examples{
present_genes_SCT(obj, gene_list, genecolumn)


}
