% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFARI_gene.R
\docType{data}
\name{SFARI_genes}
\alias{SFARI_genes}
\title{SFARI gene data}
\format{
A data frame with 1140 genes in ENSG and gene symbol format.
}
\source{
https://gene.sfari.org/database/human-gene/
}
\usage{
data(SFARI_genes)
}
\description{
Data from the SFARI gene database.
}
\examples{

data(SFARI_genes)
}
\keyword{datasets}
