% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.r
\name{correlation}
\alias{correlation}
\title{Finds correlation between a gene of interest and all other genes in a Seurat object}
\usage{
correlation(df, goi, gene_list)
}
\arguments{
\item{obj}{Seurat object. gene_list Gene list. goi Gene of interest.}
}
\value{
A vector of present genes
}
\description{
Finds correlation between a gene of interest and all other genes in a Seurat object
}
\examples{
correlation(df, goi, gene_list)


}
